@@ -11,15 +11,15 @@ static def generateVersionCode() {
 }
 
 android {
-    compileSdkVersion 19
+    compileSdkVersion 28
     buildToolsVersion "26.0.0"
     defaultConfig {
         applicationId "later.com.linkinbio"
         minSdkVersion 15
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode generateVersionCode()
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -30,13 +30,15 @@ android {
 }
 
 def retrofitVersion = '2.7.1'
+def lifeCycleExtensionsVersion = '1.1.1'
+def androidxVersion = '1.1.0'
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    implementation 'com.android.support:appcompat-v7:26.0.0-beta2'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
     testImplementation 'junit:junit:4.12'
 
     // Retrofit 2
@@ -44,6 +46,16 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
     implementation "com.squareup.retrofit2:adapter-rxjava2:$retrofitVersion"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+
+    implementation "io.reactivex.rxjava2:rxandroid:2.1.1"
+
+    // Android Lifecycle Extensions
+    implementation "android.arch.lifecycle:extensions:$lifeCycleExtensionsVersion"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    implementation "androidx.activity:activity-ktx:$androidxVersion"
+
+
 }
 repositories {
     mavenCentral()
