@@ -1,6 +1,6 @@
 package later.com.linkinbio.model
 
 data class LinkinbioPost(
-    val image_url: String,
-    val link_url: String,
+    val image_url: String?,
+    val link_url: String?,
 )
